GRZip: port the WFC coder; share the arithmetic core with MTF
210/210 differential cases byte-identical to the C decoder -- 21 inputs across
ten modes covering all four transform/coder combinations (BWT/ST4 x MTF/WFC)
with LZP off and on. Every stage of GRZip's block decoder is now ported and
verified.

MTF_Ari.c and WFC_Ari.c are near-duplicates: identical range coder, identical
model stack, identical run-length ladder. They differ in exactly one step --
turning a decoded rank into a character. So rather than transcribe those ~150
lines a second time, they moved to grzip/ari.rs behind a `SymbolList` trait,
with mtf_ari.rs and wfc_ari.rs supplying only the list discipline. The MTF path
was re-run through the harness after the move to confirm no regression, and the
WFC coder then worked first time -- which is the argument for the refactor: the
duplicated half was the part already proven correct.

What WFC does differently: move-to-front promotes a character to the head on
every use, so one appearance of a rare byte costs every other byte a rank. WFC
keeps a weight per character and sorts by it -- the current character gains
WFC_Val0, and the characters 1, 2, 4 ... 2048 symbols back each lose a
decreasing amount. Those twelve decrements sum to exactly WFC_Val0, so total
weight is conserved. Hence the history buffer: one byte per decoded *symbol*,
not per output byte.

Two traps recorded in the source:

  * The C's parameter order differs between the two functions.
    GRZip_MTF_Ari_Decode takes (Input, Size = compressed length, Output,
    OutSize); GRZip_WFC_Ari_Decode takes (Input, Size = *decoded* length,
    Output, InSize = compressed length). The second and fourth arguments swap
    meaning between two otherwise near-identical functions.
  * CharWeight[256] = -1 is a sentinel terminating the insertion walk, which
    only works while every real weight stays above -1. The walk is bounded at
    the sentinel index here rather than trusting that.

The harness now covers all ten modes; the WFC-rejection assertion it carried
while WFC was stubbed is gone, replaced by real coverage.
